Extracting the request creator & building a writer
Extracted the code the generates the Parameters in a CRD Operation
request and put it in a stand alone class. Provided ids and references
as well as brought the sample data in line with what was discussed at
the Cambridge work group.

Created a small application that will write out the request as JSON and
XML.