Minor web cleanup and dev-tools
Added a header to the application. Pulled the header and data table out
to fragments. Put in some placeholder text and provided better variable
names in the HomeController.

Also added spring-boot-devtools to the project. This allows auto
reloading of changes if you run `gradle build --continuous` in a
separate process.