Adding a code_system column to the database
Codes belong to code systems. We have sample data that belongs to both
CPT and HCPCS. This change adds a column to the database to store the
code system. Updates the sample data with the correct code system. It
updates the CoverageRequirementRule model and the web UI. It also
updates the request-generator so that values there now have a code
system and that value is no longer hard coded to HCPCS.