Small fixes for RequestBuilder
Updates the URL for R4
Due to the way prefetch works in CDS Hooks, the DeviceRequest is present
in both the context as well as the prefetch. Previous changes fixed the
code system in the prefetch, but not in the context. This change fixes
the code system in the DeviceRequest in the context.