Warning card and remove devtools
If the services reach the end of their processing, but the code in the
service has not added any cards, a warning card will be generated
stating that the request didn't have enough information. This typically
happens when the patient is null (broken reference) or it can't find a
CodeableConcept in the passed in request resource.

This also removes the spring devtools which was causing the server to
reload on each request.