Removing the default value from DropdownState
If a default value is set, the UI in RequestBuilder becomes confusing.
Both state drop downs will be populated, but since there is no state
change, updateCB isn't called so the component state in RequestBuilder
does not have have a patient or provider state set. If you fill out
everything else, it will submit the request with no states. This makes
you pick a state.